+++ src/app/app.module.ts
@@ -5,19 +5,22 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { TimerPageComponent } from './timer-page/timer-page.component';
-import { MainPageComponent } from './main-page/main-page.component';
 
 import { MatButtonModule } from '@angular/material/button';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCardModule } from '@angular/material/card';
 
+import { TimerPageComponent } from './timer-page/timer-page.component';
+import { MainPageComponent } from './main-page/main-page.component';
+import { TimerComponent } from './timer/timer.component';
+
 @NgModule({
   declarations: [
     AppComponent,
     TimerPageComponent,
-    MainPageComponent
+    MainPageComponent,
+    TimerComponent
   ],
   imports: [
     FormsModule,